Move the `/connection/data` endpoint to the Connection package.
