+++ projects/plugins/jetpack/_inc/lib/class.core-rest-api-endpoints.php
@@ -162,17 +162,6 @@ class Jetpack_Core_Json_Api_Endpoints {
 			)
 		);
 
-		// Get current user connection data
-		register_rest_route(
-			'jetpack/v4',
-			'/connection/data',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => __CLASS__ . '::get_user_connection_data',
-				'permission_callback' => __CLASS__ . '::get_user_connection_data_permission_callback',
-			)
-		);
-
 		// Current user: get or set tracking settings.
 		register_rest_route(
 			'jetpack/v4',
@@ -1246,15 +1235,13 @@ class Jetpack_Core_Json_Api_Endpoints {
 	 * Only those who can connect.
 	 *
 	 * @since 4.3.0
+	 * @deprecated 10.0.0
 	 *
 	 * @return bool|WP_Error True if user is able to unlink.
 	 */
 	public static function get_user_connection_data_permission_callback() {
-		if ( current_user_can( 'jetpack_connect_user' ) ) {
-			return true;
-		}
-
-		return new WP_Error( 'invalid_user_permission_user_connection_data', self::$user_permissions_error_msg, array( 'status' => rest_authorization_required_code() ) );
+		_deprecated_function( __METHOD__, 'jetpack-10.0.0', 'Automattic\\Jetpack\\Connection\\REST_Connector::get_user_connection_data_permission_callback' );
+		return REST_Connector::get_user_connection_data_permission_callback();
 	}
 
 	/**
@@ -1766,22 +1753,13 @@ class Jetpack_Core_Json_Api_Endpoints {
 	 * Information about the current user.
 	 *
 	 * @since 4.3.0
-	 *
-	 * @param WP_REST_Request $request The request sent to the WP REST API.
+	 * @deprecated 10.0.0
 	 *
 	 * @return object
 	 */
 	public static function get_user_connection_data() {
-		require_once JETPACK__PLUGIN_DIR . '_inc/lib/admin-pages/class.jetpack-react-page.php';
-
-		$connection_owner   = ( new Connection_Manager() )->get_connection_owner();
-		$owner_display_name = false === $connection_owner ? null : $connection_owner->data->display_name;
-
-		$response = array(
-			'currentUser'     => jetpack_current_user_data(),
-			'connectionOwner' => $owner_display_name,
-		);
-		return rest_ensure_response( $response );
+		_deprecated_function( __METHOD__, 'jetpack-10.0.0', 'Automattic\\Jetpack\\Connection\\REST_Connector::get_user_connection_data' );
+		return REST_Connector::get_user_connection_data();
 	}
 
 	/**
